Implemented both parts, untested

+++ kernel/stat.h
@@ -4,8 +4,8 @@
 
 struct stat {
   int dev;     // File system's disk device
-  uint ino;    // Inode number
+  int ino;    // Inode number
   short type;  // Type of file
   short nlink; // Number of links to file
-  uint64 size; // Size of file in bytes
+  int size; // Size of file in bytes
 };
